requirements.txt
textarena-sdk
numpy
pandas
pyyaml
regex
torch
transformers